+++ src/api/adapter/standalone.rs
@@ -8,9 +8,7 @@ use isahc::Request;
 use crate::api::adapter::{
     build_http_client, execute_request, http_ping, InternalHttpClient, MockServerAdapter,
 };
-use crate::data::{
-    ActiveMock, ClosestMatch, MockDefinition, MockIdentification, RequestRequirements,
-};
+use crate::common::data::{ActiveMock, ClosestMatch, MockDefinition, MockRef, RequestRequirements};
 
 #[derive(Debug)]
 pub struct RemoteMockServerAdapter {
@@ -50,7 +48,7 @@ impl MockServerAdapter for RemoteMockServerAdapter {
         &self.addr
     }
 
-    async fn create_mock(&self, mock: &MockDefinition) -> Result<MockIdentification, String> {
+    async fn create_mock(&self, mock: &MockDefinition) -> Result<MockRef, String> {
         // Check if the request can be sent via HTTP
         self.validate_mock(mock).expect("Cannot create mock");
 
@@ -83,7 +81,7 @@ impl MockServerAdapter for RemoteMockServerAdapter {
         }
 
         // Create response object
-        let response: serde_json::Result<MockIdentification> = serde_json::from_str(&body);
+        let response: serde_json::Result<MockRef> = serde_json::from_str(&body);
         if let Err(err) = response {
             return Err(format!("Cannot deserialize mock server response: {}", err));
         }