+++ src/api/mod.rs
@@ -5,7 +5,10 @@ pub use adapter::{
     local::LocalMockServerAdapter, standalone::RemoteMockServerAdapter, Method, MockServerAdapter,
     Regex,
 };
+pub use mock::{Mock, MockExt};
+pub use server::MockServer;
 
-pub use mock::{MockRef, MockRefExt};
 mod adapter;
 mod mock;
+mod server;
+pub mod spec;