+++ src/server/web/routes.rs
@@ -3,7 +3,10 @@ use std::collections::BTreeMap;
 use qstring::QString;
 use serde::Serialize;
 
-use crate::data::*;
+use crate::common::data::{
+    ErrorResponse, HttpMockRequest, MockDefinition, MockRef, MockServerHttpResponse,
+    RequestRequirements,
+};
 use crate::server::web::handlers;
 use crate::server::{MockServerState, ServerRequestHeader, ServerResponse};
 use std::time::Instant;
@@ -27,7 +30,7 @@ pub(crate) fn add(state: &MockServerState, body: Vec<u8>) -> Result<ServerRespon
 
     match result {
         Err(e) => create_json_response(500, None, ErrorResponse::new(&e)),
-        Ok(mock_id) => create_json_response(201, None, MockIdentification { mock_id }),
+        Ok(mock_id) => create_json_response(201, None, MockRef { mock_id }),
     }
 }